--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -177,6 +177,7 @@ set(LIGHTSTEP_SRCS src/common/utility.cpp
                    src/common/random_traverser.cpp
                    src/common/serialization.cpp
                    src/common/serialization_chain.cpp
+                   src/common/timestamp.cpp
                    src/recorder/report_builder.cpp
                    src/recorder/auto_recorder.cpp
                    src/recorder/fork_aware_recorder.cpp