--- a/src/common/timestamp.h
+++ b/src/common/timestamp.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <chrono>
+#include <cstdint>
+
+namespace lightstep {
+int64_t ComputeSystemSteadyTimestampDelta() noexcept;
+
+std::chrono::system_clock::time_point ToSystemTimestamp(
+    int64_t timestamp_delta,
+    std::chrono::steady_clock::time_point steady_time) noexcept;
+
+std::chrono::steady_clock::time_point ToSteadyTimestamp(
+    int64_t timestamp_delta,
+    std::chrono::system_clock::time_point system_time) noexcept;
+}  // namespace lightstep