--- a/src/recorder/BUILD
+++ b/src/recorder/BUILD
@@ -36,6 +36,7 @@ lightstep_cc_library(
     ],
     deps = [
         "//src/common:serialization_chain_lib",
+        "//src/common:timestamp_lib",
         "//lightstep-tracer-common:collector_proto_cc",
         "//include/lightstep:tracer_interface",
     ],