--- a/src/recorder/stream_recorder/stream_recorder_impl.cpp
+++ b/src/recorder/stream_recorder/stream_recorder_impl.cpp
@@ -76,6 +76,13 @@ void StreamRecorderImpl::Poll() noexcept {
 }
 
 //--------------------------------------------------------------------------------------------------
+// ComputeTimestampDelta
+//--------------------------------------------------------------------------------------------------
+void StreamRecorderImpl::ComputeTimestampDelta() noexcept {
+  timestamp_delta_ = ComputeSystemSteadyTimestampDelta();
+}
+
+//--------------------------------------------------------------------------------------------------
 // Flush
 //--------------------------------------------------------------------------------------------------
 void StreamRecorderImpl::Flush() noexcept try {