--- a/src/recorder/stream_recorder/stream_recorder_impl.h
+++ b/src/recorder/stream_recorder/stream_recorder_impl.h
@@ -4,6 +4,7 @@
 #include <thread>
 
 #include "common/noncopyable.h"
+#include "common/timestamp.h"
 #include "network/event_base.h"
 #include "network/timer_event.h"
 #include "recorder/stream_recorder/satellite_streamer.h"
@@ -24,22 +25,28 @@ class StreamRecorderImpl : private Noncopyable {
 
   ~StreamRecorderImpl() noexcept;
 
+  int64_t timestamp_delta() const noexcept { return timestamp_delta_; }
+
  private:
   StreamRecorder& stream_recorder_;
 
   EventBase event_base_;
   size_t early_flush_marker_;
   TimerEvent poll_timer_;
+  TimerEvent timestamp_delta_timer_;
   TimerEvent flush_timer_;
 
   SatelliteStreamer streamer_;
 
   std::thread thread_;
   std::atomic<bool> exit_{false};
+  std::atomic<int64_t> timestamp_delta_{ComputeSystemSteadyTimestampDelta()};
 
   void Run() noexcept;
 
   void Poll() noexcept;
+
+  void ComputeTimestampDelta() noexcept;
 
   void Flush() noexcept;
 };