--- a/src/recorder/stream_recorder/stream_recorder_impl2.h
+++ b/src/recorder/stream_recorder/stream_recorder_impl2.h
@@ -49,7 +49,7 @@ class StreamRecorderImpl2 : private Noncopyable {
 
   void Poll() noexcept;
 
-  void ComputeTimestampDelta() noexcept;
+  void RefreshTimestampDelta() noexcept;
 
   void Flush() noexcept;
 };