--- a/src/tracer/utility.cpp
+++ b/src/tracer/utility.cpp
@@ -14,27 +14,28 @@ bool is_sampled(const opentracing::Value& value) noexcept {
 // ComputeStartTimestamps
 //------------------------------------------------------------------------------
 std::tuple<SystemTime, SteadyTime> ComputeStartTimestamps(
-    const SystemTime& start_system_timestamp,
+    const Recorder& recorder, const SystemTime& start_system_timestamp,
     const SteadyTime& start_steady_timestamp) noexcept {
   // If neither the system nor steady timestamps are set, get the tme from the
   // respective clocks; otherwise, use the set timestamp to initialize the
   // other.
   if (start_system_timestamp == SystemTime() &&
       start_steady_timestamp == SteadyTime()) {
-    return std::tuple<SystemTime, SteadyTime>{SystemClock::now(),
-                                              SteadyClock::now()};
-    /* return std::tuple<SystemTime, SteadyTime>{SystemTime{}, */
-    /*                                           SteadyClock::now()}; */
+    auto steady_now = SteadyClock::now();
+    return std::tuple<SystemTime, SteadyTime>{
+        recorder.ComputeCurrentSystemTimestamp(steady_now), steady_now};
   }
   if (start_system_timestamp == SystemTime()) {
+    auto timestamp_delta = recorder.ComputeSystemSteadyTimestampDelta();
     return std::tuple<SystemTime, SteadyTime>{
-        opentracing::convert_time_point<SystemClock>(start_steady_timestamp),
+        ToSystemTimestamp(timestamp_delta, start_steady_timestamp),
         start_steady_timestamp};
   }
   if (start_steady_timestamp == SteadyTime()) {
+    auto timestamp_delta = recorder.ComputeSystemSteadyTimestampDelta();
     return std::tuple<SystemTime, SteadyTime>{
         start_system_timestamp,
-        opentracing::convert_time_point<SteadyClock>(start_system_timestamp)};
+        ToSteadyTimestamp(timestamp_delta, start_system_timestamp)};
   }
   return std::tuple<SystemTime, SteadyTime>{start_system_timestamp,
                                             start_steady_timestamp};