Avoids the cost of timestamping twice when starting a span by caching the system-steady delta in the recorder.

These are the benchmark times I got running on my machine before the change:
```
-----------------------------------------------------------------------------
Benchmark                                      Time           CPU Iterations
-----------------------------------------------------------------------------
BM_SpanCreation/rpc                        18012 ns      17830 ns      33544
BM_SpanCreation/stream                     18525 ns      18497 ns      36456
BM_SpanCreation/stream2                    18736 ns      18706 ns      37213
```
and after the change:
```
-----------------------------------------------------------------------------
Benchmark                                      Time           CPU Iterations
-----------------------------------------------------------------------------
BM_SpanCreation/rpc                        18401 ns      18077 ns      35548
BM_SpanCreation/stream                     13388 ns      13345 ns      49015
BM_SpanCreation/stream2                    14892 ns      14837 ns      34336
```
(this change only affects the stream times)